+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 # Upcoming
 
 ### 🔄 Changed
+- Exposed `SearchResultsView` as a public component
+- `isSearching` property in the `ChatChannelListViewModel` is now public
 
 # [4.28.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.28.0)
 _February 28, 2023_